+++ src/main/java/study/sharedcalendar/service/UserService.java
@@ -2,10 +2,7 @@ package study.sharedcalendar.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import study.sharedcalendar.dto.LoginReq;
-import study.sharedcalendar.dto.LoginRes;
 import study.sharedcalendar.dto.SignUpReq;
-import study.sharedcalendar.exception.AuthorizationException;
 import study.sharedcalendar.exception.DuplicateException;
 import study.sharedcalendar.mapper.UserMapper;
 
@@ -16,7 +13,6 @@ import static study.constant.ErrorCode.*;
 public class UserService {
     private final UserMapper userMapper;
     private final EncryptionService encryptionService;
-    private final SessionService sessionService;
 
     public void signUp(SignUpReq signUpReq) {
         String encryptedPassword = encryptionService.encrypt(signUpReq.getPassword());
@@ -39,16 +35,4 @@ public class UserService {
         return userMapper.userIdExist(userId);
     }
 
-    public void login(LoginReq loginReq) {
-        LoginRes loginRes = userMapper.findLoginUser(loginReq);
-
-        if (loginRes == null) {
-            throw new NullPointerException("일치하는 아이디가 없습니다.");
-        }
-        if (!encryptionService.isMatch(loginReq.getPassword(), loginRes.getPassword())) {
-            throw new AuthorizationException(NO_MATCHING_USER_PASSWORD);
-        }
-
-        sessionService.setLoginSession(loginReq.getUserId());
-    }
 }
